routes: migrate gates + tasks domains (15 routes)
Move to:
- src/server/routes/tasks.ts (7): tasks CRUD + assign/transition + nested
  /goals/:id/tasks
- src/server/routes/gates.ts (8): gates list/detail/inspect/signals/content/
  workflow-context, verifications/active, gates/:gateId/cancel-verification

POST /api/goals/:id/gates/:gateId/signal stays in server.ts — verification
harness + cascade reset coupling worth its own follow-up commit.

Drops now-unused inner closures getTaskManagerForGoal, getTaskManagerForTask
and TaskState/VALID_TASK_STATES imports.

Net: server.ts -462 lines (5409 → 4947). No behaviour change.
