+++ src/server/server.ts
@@ -29,7 +29,6 @@ import { recordElapsed } from "./agent/profiling.js";
 
 import { initSkillSidecarDir } from "./skills/skill-sidecar.js";
 import { buildActivationHeader } from "./skills/skill-manifest.js";
-import type { TaskState } from "./agent/task-store.js";
 import { TaskManager } from "./agent/task-manager.js";
 import { TaskStore } from "./agent/task-store.js";
 import { BgProcessManager } from "./agent/bg-process-manager.js";
@@ -98,8 +97,6 @@ import {
 	writeProposalFile,
 	type ProposalType,
 } from "./proposals/proposal-files.js";
-
-const VALID_TASK_STATES = new Set<string>(["todo", "in-progress", "blocked", "complete", "skipped"]);
 
 const execAsync = promisify(exec);
 const execFileAsync = promisify(execFileCb);
@@ -1106,30 +1103,6 @@ async function handleApiRoute(
 		const ctx = projectContextManager.getContextForGoal(goalId);
 		if (!ctx) throw new Error(`Goal "${goalId}" not found in any project`);
 		return ctx.goalManager;
-	}
-
-	/** Get a TaskManager for the project that owns the given goal. Throws if not found. */
-	const taskManagerCache = new Map<string, TaskManager>();
-	function getTaskManagerForGoal(goalId: string): TaskManager {
-		const ctx = projectContextManager.getContextForGoal(goalId);
-		if (!ctx) throw new Error(`Goal "${goalId}" not found in any project`);
-		const projectId = ctx.project.id;
-		let tm = taskManagerCache.get(projectId);
-		if (!tm) {
-			tm = new TaskManager(ctx.taskStore);
-			taskManagerCache.set(projectId, tm);
-		}
-		return tm;
-	}
-
-	/** Get a TaskManager for a task by looking up which goal it belongs to. Throws if not found. */
-	function getTaskManagerForTask(taskId: string): TaskManager {
-		// Search all project contexts for the task
-		for (const ctx of projectContextManager.all()) {
-			const task = ctx.taskStore.get(taskId);
-			if (task) return getTaskManagerForGoal(task.goalId);
-		}
-		throw new Error(`Task "${taskId}" not found in any project`);
 	}
 
 	// ── Project Detection & Browse ────────────────────────────────────
@@ -2472,233 +2445,6 @@ async function handleApiRoute(
 
 	// ── Task endpoints ─────────────────────────────────────────────
 
-	// GET /api/goals/:goalId/tasks — list tasks for a goal
-	const goalTasksMatch = url.pathname.match(/^\/api\/goals\/([^/]+)\/tasks$/);
-	if (goalTasksMatch && req.method === "GET") {
-		const tasks = getTaskManagerForGoal(goalTasksMatch[1]).getTasksForGoal(goalTasksMatch[1]);
-		if (url.searchParams.get("view") === "summary") {
-			const slim = tasks.map(t => ({
-				id: t.id,
-				title: t.title,
-				type: t.type,
-				state: t.state,
-				assignedSessionId: t.assignedSessionId,
-				branch: t.branch,
-				headSha: t.headSha,
-				workflowGateId: t.workflowGateId,
-				dependsOn: t.dependsOn || [],
-			}));
-			json({ tasks: slim });
-			return;
-		}
-		json({ tasks });
-		return;
-	}
-
-	// POST /api/goals/:goalId/tasks — create a task
-	if (goalTasksMatch && req.method === "POST") {
-		const goalId = goalTasksMatch[1];
-		const goal = getGoalAcrossProjects(goalId);
-		if (!goal) { json({ error: "Goal not found" }, 404); return; }
-		if (goal.archived) { json({ error: "Goal is archived" }, 409); return; }
-
-		const body = await readBody(req);
-		const title = body?.title;
-		const type = body?.type;
-		if (!title || typeof title !== "string") {
-			json({ error: "Missing title" }, 400);
-			return;
-		}
-		if (!type || typeof type !== "string") {
-			json({ error: "Missing type" }, 400);
-			return;
-		}
-		try {
-			const task = getTaskManagerForGoal(goalId).createTask(goalId, title, type, {
-				parentTaskId: body.parentTaskId,
-				spec: body.spec,
-				dependsOn: body.dependsOn,
-				workflowGateId: typeof body.workflowGateId === "string" ? body.workflowGateId : undefined,
-				inputGateIds: Array.isArray(body.inputGateIds) ? body.inputGateIds as string[] : undefined,
-			});
-			json(task, 201);
-		} catch (err: any) {
-			jsonError(400, err);
-		}
-		return;
-	}
-
-	// ── Gate endpoints ─────────────────────────────────────────────
-
-	// GET /api/goals/:goalId/gates — list gates for a goal
-	const goalGatesMatch = url.pathname.match(/^\/api\/goals\/([^/]+)\/gates$/);
-	if (goalGatesMatch && req.method === "GET") {
-		const goalId = goalGatesMatch[1];
-		const goal = getGoalAcrossProjects(goalId);
-		if (!goal) { json({ error: "Goal not found" }, 404); return; }
-		const gateCtx = projectContextManager.getContextForGoal(goalId);
-		if (!gateCtx) { json({ error: "Goal not found in any project" }, 404); return; }
-		const gateStore = gateCtx.gateStore;
-		const gates = gateStore.getGatesForGoal(goalId);
-		// Enrich with workflow gate definitions
-		const enriched = gates.map(g => {
-			const def = goal.workflow?.gates.find(wg => wg.id === g.gateId);
-			return { ...g, name: def?.name, dependsOn: def?.dependsOn, content: def?.content, injectDownstream: def?.injectDownstream, metadata: def?.metadata || g.currentMetadata, signalCount: g.signals.length };
-		});
-		if (url.searchParams.get("view") === "summary") {
-			const slim = enriched.map(g => {
-				const base: Record<string, unknown> = {
-					gateId: g.gateId,
-					name: g.name,
-					status: g.status,
-					dependsOn: g.dependsOn || [],
-					signalCount: g.signalCount,
-				};
-				if (g.signals.length > 0) base.updatedAt = g.updatedAt;
-				if (g.status === "failed") {
-					const latest = g.signals[g.signals.length - 1];
-					if (latest?.verification?.steps) {
-						base.failedSteps = latest.verification.steps
-							.filter((s: any) => !s.passed && !s.skipped)
-							.map((s: any) => s.name);
-					}
-				}
-				return base;
-			});
-			json({ gates: slim });
-			return;
-		}
-		json({ gates: enriched });
-		return;
-	}
-
-	// GET /api/goals/:goalId/gates/:gateId — gate detail
-	const gateDetailMatch = url.pathname.match(/^\/api\/goals\/([^/]+)\/gates\/([^/]+)$/);
-	if (gateDetailMatch && req.method === "GET") {
-		const [, goalId, gateId] = gateDetailMatch;
-		const gateDetailCtx = projectContextManager.getContextForGoal(goalId);
-		if (!gateDetailCtx) { json({ error: "Goal not found in any project" }, 404); return; }
-		const gateStore = gateDetailCtx.gateStore;
-		const gate = gateStore.getGate(goalId, gateId);
-		if (!gate) { json({ error: "Gate not found" }, 404); return; }
-		const goal = getGoalAcrossProjects(goalId);
-		const def = goal?.workflow?.gates.find(wg => wg.id === gateId);
-		if (url.searchParams.get("view") === "summary") {
-			const latestSignal = gate.signals[gate.signals.length - 1];
-			const MAX_TAIL_LINES = 40;
-			const slim: Record<string, unknown> = {
-				gateId: gate.gateId,
-				name: def?.name,
-				status: gate.status,
-				dependsOn: def?.dependsOn || [],
-				signalCount: gate.signals.length,
-				updatedAt: gate.updatedAt,
-				hasContent: !!gate.currentContent,
-				contentLength: gate.currentContent?.length || 0,
-			};
-			if (gate.currentMetadata) slim.currentMetadata = gate.currentMetadata;
-			if (latestSignal) {
-				slim.latestSignal = {
-					id: latestSignal.id,
-					sessionId: latestSignal.sessionId,
-					timestamp: latestSignal.timestamp,
-					commitSha: latestSignal.commitSha,
-					verification: latestSignal.verification ? {
-						status: latestSignal.verification.status,
-						steps: latestSignal.verification.steps.map(s => {
-							const base: Record<string, unknown> = { name: s.name, passed: s.passed };
-							if (!s.passed && !s.skipped && s.output) {
-								const lines = s.output.split("\n");
-								base.output = lines.length > MAX_TAIL_LINES
-									? lines.slice(-MAX_TAIL_LINES).join("\n")
-									: s.output;
-							}
-							return base;
-						}),
-					} : undefined,
-				};
-			}
-			json(slim);
-			return;
-		}
-		json({ ...gate, name: def?.name, dependsOn: def?.dependsOn, content: def?.content, injectDownstream: def?.injectDownstream });
-		return;
-	}
-
-	// GET /api/goals/:goalId/gates/:gateId/inspect — scoped gate data retrieval
-	const gateInspectMatch = url.pathname.match(/^\/api\/goals\/([^/]+)\/gates\/([^/]+)\/inspect$/);
-	if (gateInspectMatch && req.method === "GET") {
-		const [, goalId, gateId] = gateInspectMatch;
-		const ctx = projectContextManager.getContextForGoal(goalId);
-		if (!ctx) { json({ error: "Goal not found" }, 404); return; }
-		const gate = ctx.gateStore.getGate(goalId, gateId);
-		if (!gate) { json({ error: "Gate not found" }, 404); return; }
-
-		const section = url.searchParams.get("section");
-		if (!section || !["content", "verification", "signals"].includes(section)) {
-			json({ error: "section query parameter is required: 'content', 'verification', or 'signals'" }, 400);
-			return;
-		}
-
-		const resolveSignal = () => {
-			const idxStr = url.searchParams.get("signal_index");
-			let idx = idxStr !== null ? parseInt(idxStr, 10) : -1;
-			if (isNaN(idx)) idx = -1;
-			if (idx < 0) idx = gate.signals.length + idx;
-			if (idx < 0 || idx >= gate.signals.length) return null;
-			return { signal: gate.signals[idx], index: idx };
-		};
-
-		if (section === "content") {
-			const resolved = resolveSignal();
-			if (!resolved) { json({ error: "Signal not found" }, 404); return; }
-			json({
-				gateId, section: "content",
-				signalIndex: resolved.index,
-				signalId: resolved.signal.id,
-				text: resolved.signal.content || null,
-			});
-			return;
-		}
-
-		if (section === "verification") {
-			const resolved = resolveSignal();
-			if (!resolved) { json({ error: "Signal not found" }, 404); return; }
-			const v = resolved.signal.verification;
-			json({
-				gateId, section: "verification",
-				signalIndex: resolved.index,
-				signalId: resolved.signal.id,
-				steps: v ? v.steps.map(s => ({
-					name: s.name,
-					type: s.type,
-					passed: s.passed,
-					skipped: s.skipped || undefined,
-					duration_ms: s.duration_ms,
-					output: s.output,
-				})) : [],
-			});
-			return;
-		}
-
-		if (section === "signals") {
-			json({
-				gateId, section: "signals",
-				signals: gate.signals.map((s, i) => ({
-					index: i,
-					id: s.id,
-					timestamp: s.timestamp,
-					sessionId: s.sessionId,
-					commitSha: s.commitSha,
-					verdict: s.verification?.status || "running",
-					hasContent: !!s.content,
-					metadataKeys: s.metadata ? Object.keys(s.metadata) : [],
-				})),
-			});
-			return;
-		}
-	}
-
 	// POST /api/goals/:goalId/gates/:gateId/signal — signal a gate
 	const gateSignalMatch = url.pathname.match(/^\/api\/goals\/([^/]+)\/gates\/([^/]+)\/signal$/);
 	if (gateSignalMatch && req.method === "POST") {
@@ -2883,214 +2629,6 @@ async function handleApiRoute(
 
 		const verifySteps = (gateDef.verify || []).map((s: any) => ({ name: s.name, type: s.type }));
 		json({ signal: { id: signal.id, gateId, goalId, status: "running", steps: verifySteps } }, 201);
-		return;
-	}
-
-	// GET /api/goals/:goalId/gates/:gateId/signals — signal history
-	const gateSignalsMatch = url.pathname.match(/^\/api\/goals\/([^/]+)\/gates\/([^/]+)\/signals$/);
-	if (gateSignalsMatch && req.method === "GET") {
-		const [, goalId, gateId] = gateSignalsMatch;
-		const gateSignalsCtx = projectContextManager.getContextForGoal(goalId);
-		if (!gateSignalsCtx) { json({ error: "Goal not found in any project" }, 404); return; }
-		const gateStore = gateSignalsCtx.gateStore;
-		const gate = gateStore.getGate(goalId, gateId);
-		if (!gate) { json({ error: "Gate not found" }, 404); return; }
-		json({ signals: gate.signals });
-		return;
-	}
-
-	// GET /api/goals/:goalId/verifications/active — get in-flight verification state
-	const activeVerifMatch = url.pathname.match(/^\/api\/goals\/([^/]+)\/verifications\/active$/);
-	if (activeVerifMatch && req.method === "GET") {
-		const [, goalId] = activeVerifMatch;
-		const active = verificationHarness.getActiveVerifications(goalId);
-		json({ verifications: active });
-		return;
-	}
-
-	// POST /api/goals/:goalId/gates/:gateId/cancel-verification — cancel a stuck verification
-	const cancelVerifMatch = url.pathname.match(/^\/api\/goals\/([^/]+)\/gates\/([^/]+)\/cancel-verification$/);
-	if (cancelVerifMatch && req.method === "POST") {
-		const [, goalId, gateId] = cancelVerifMatch;
-		const goal = getGoalAcrossProjects(goalId);
-		if (!goal) { json({ error: "Goal not found" }, 404); return; }
-		if (goal.archived) { json({ error: "Goal is archived" }, 409); return; }
-		if (goal.state === "shelved") { json({ error: "Goal is shelved" }, 400); return; }
-
-		const activeVers = verificationHarness.getActiveVerifications(goalId);
-		const running = activeVers.find(v => v.gateId === gateId && v.overallStatus === "running");
-		if (!running) {
-			json({ cancelled: false, message: "No running verification to cancel" }, 200);
-			return;
-		}
-
-		await verificationHarness.cancelStaleVerifications(goalId, gateId);
-		// Explicit user cancel: also update gate status to "failed"
-		const cancelCtx = projectContextManager.getContextForGoal(goalId);
-		if (cancelCtx) cancelCtx.gateStore.updateGateStatus(goalId, gateId, "failed");
-		json({ cancelled: true }, 200);
-		return;
-	}
-
-	// GET /api/goals/:goalId/gates/:gateId/content — gate content
-	const gateContentMatch = url.pathname.match(/^\/api\/goals\/([^/]+)\/gates\/([^/]+)\/content$/);
-	if (gateContentMatch && req.method === "GET") {
-		const [, goalId, gateId] = gateContentMatch;
-		const gateContentCtx = projectContextManager.getContextForGoal(goalId);
-		if (!gateContentCtx) { json({ error: "Goal not found in any project" }, 404); return; }
-		const gateStore = gateContentCtx.gateStore;
-		const gate = gateStore.getGate(goalId, gateId);
-		if (!gate) { json({ error: "Gate not found" }, 404); return; }
-		json({ content: gate.currentContent, version: gate.currentContentVersion });
-		return;
-	}
-
-	// GET /api/goals/:goalId/workflow-context/:gateId — get dependency context for a gate
-	const workflowContextMatch = url.pathname.match(/^\/api\/goals\/([^/]+)\/workflow-context\/([^/]+)$/);
-	if (workflowContextMatch && req.method === "GET") {
-		const goalId = workflowContextMatch[1];
-		const gateId = workflowContextMatch[2];
-		const goal = getGoalAcrossProjects(goalId);
-		if (!goal) { json({ error: "Goal not found" }, 404); return; }
-		if (!goal.workflow) { json({ error: "Goal has no workflow" }, 404); return; }
-		const gateDef = goal.workflow.gates.find(g => g.id === gateId);
-		if (!gateDef) { json({ error: "Gate not found" }, 404); return; }
-
-		const context = teamManager.buildDependencyContext(goalId, gateId);
-		json({ context, gate: gateDef });
-		return;
-	}
-
-	// Routes with task :id parameter
-	const taskMatch = url.pathname.match(/^\/api\/tasks\/([^/]+)$/);
-	if (taskMatch) {
-		const id = taskMatch[1];
-
-		// GET /api/tasks/:id
-		if (req.method === "GET") {
-			try {
-				const task = getTaskManagerForTask(id).getTask(id);
-				if (!task) { json({ error: "Task not found" }, 404); return; }
-				json(task);
-			} catch {
-				json({ error: "Task not found" }, 404);
-			}
-			return;
-		}
-
-		// PUT /api/tasks/:id
-		if (req.method === "PUT") {
-			const body = await readBody(req);
-			if (!body) { json({ error: "Missing body" }, 400); return; }
-			try {
-				const tm = getTaskManagerForTask(id);
-				const task = tm.getTask(id);
-				const prevState = task?.state;
-				const ok = tm.updateTask(id, {
-					title: body.title,
-					spec: body.spec,
-					state: body.state,
-					assignedSessionId: body.assignedSessionId,
-					dependsOn: body.dependsOn,
-					workflowGateId: typeof body.workflowGateId === "string" ? body.workflowGateId : undefined,
-					inputGateIds: Array.isArray(body.inputGateIds) ? body.inputGateIds as string[] : undefined,
-					headSha: typeof body.headSha === "string" ? body.headSha : undefined,
-					baseSha: typeof body.baseSha === "string" ? body.baseSha : undefined,
-					branch: typeof body.branch === "string" ? body.branch : undefined,
-					resultSummary: typeof body.resultSummary === "string" ? body.resultSummary : undefined,
-				});
-				if (!ok) { json({ error: "Task not found" }, 404); return; }
-
-				// Notify team lead when state transitions to terminal or blocked via PUT
-				if (body.state && body.state !== prevState && (body.state === "complete" || body.state === "skipped" || body.state === "blocked") && task?.goalId) {
-					teamManager.notifyTeamLeadOfTaskCompletion(task.goalId, task.title, body.state);
-				}
-
-				json({ ok: true });
-			} catch (err: any) {
-				jsonError(400, err);
-			}
-			return;
-		}
-
-		// DELETE /api/tasks/:id
-		if (req.method === "DELETE") {
-			try {
-				const ok = getTaskManagerForTask(id).deleteTask(id);
-				if (!ok) { json({ error: "Task not found" }, 404); return; }
-				json({ ok: true });
-			} catch {
-				json({ error: "Task not found" }, 404);
-			}
-			return;
-		}
-	}
-
-	// POST /api/tasks/:id/assign — assign task to session
-	const taskAssignMatch = url.pathname.match(/^\/api\/tasks\/([^/]+)\/assign$/);
-	if (taskAssignMatch && req.method === "POST") {
-		const body = await readBody(req);
-		const sessionId = body?.sessionId;
-		if (!sessionId || typeof sessionId !== "string") {
-			json({ error: "Missing sessionId" }, 400);
-			return;
-		}
-		try {
-			const taskId = taskAssignMatch[1];
-			const tm = getTaskManagerForTask(taskId);
-			const ok = tm.assignTask(taskId, sessionId);
-			if (!ok) { json({ error: "Task not found" }, 400); return; }
-
-			// Auto-populate baseSha and branch from TeamAgent record
-			const agent = teamManager.findAgentBySessionId(sessionId);
-			if (agent) {
-				const task = tm.getTask(taskId);
-				if (task) {
-					const fields: Record<string, string> = {};
-					if (agent.baseSha && !task.baseSha) fields.baseSha = agent.baseSha;
-					if (agent.branch && !task.branch) fields.branch = agent.branch;
-					if (Object.keys(fields).length) {
-						tm.updateTask(taskId, fields);
-					}
-				}
-			}
-
-			json({ ok: true });
-		} catch (err: any) {
-			jsonError(400, err);
-		}
-		return;
-	}
-
-	// POST /api/tasks/:id/transition — state transition
-	const taskTransitionMatch = url.pathname.match(/^\/api\/tasks\/([^/]+)\/transition$/);
-	if (taskTransitionMatch && req.method === "POST") {
-		const body = await readBody(req);
-		const state = body?.state;
-		if (!state || typeof state !== "string") {
-			json({ error: "Missing state" }, 400);
-			return;
-		}
-		if (!VALID_TASK_STATES.has(state)) {
-			json({ error: `Invalid task state: ${state}` }, 400);
-			return;
-		}
-		try {
-			const taskId = taskTransitionMatch[1];
-			const tm = getTaskManagerForTask(taskId);
-			const task = tm.getTask(taskId);
-			const ok = tm.transitionTask(taskId, state as TaskState);
-			if (!ok) { json({ error: "Task not found" }, 400); return; }
-
-			// Notify team lead when a task reaches a terminal or blocked state
-			if ((state === "complete" || state === "skipped" || state === "blocked") && task?.goalId) {
-				teamManager.notifyTeamLeadOfTaskCompletion(task.goalId, task.title, state);
-			}
-
-			json({ ok: true });
-		} catch (err: any) {
-			jsonError(400, err);
-		}
 		return;
 	}
